Split VoodooI2CPCILakeController into separate files
