+++ VoodooI2C/VoodooI2C/VoodooI2CController/VoodooI2CPCIController.hpp
@@ -66,15 +66,4 @@ class EXPORT VoodooI2CPCIController : public VoodooI2CController {
     void stop(IOService* provider);
 };
 
-/* Override power management for Comet Lake and newer platforms
- *
- * The members of this class are responsible for low-level interfacing with the physical PCI hardware.
- */
-
-class EXPORT VoodooI2CPCILakeController : public VoodooI2CPCIController {
-    OSDeclareDefaultStructors(VoodooI2CPCILakeController);
-
- private:
-    void configurePCI() APPLE_KEXT_OVERRIDE;
-};
 #endif /* VoodooI2CPCIController_hpp */
